Popup: add expand option.
